[refactor] 12/05 튜터님 오전 리뷰 반영
