+++ src/main/java/com/limito/payment/domain/entity/PaymentItemEntity.java
@@ -1,11 +1,19 @@
-package com.limito.payment.infrastructure.persistence.entity;
+package com.limito.payment.domain.entity;
 
 import java.util.UUID;
 
 import com.limito.common.entity.BaseEntity;
 import com.limito.payment.domain.enums.PaymentStatusEnum;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
