+++ src/main/java/com/limito/payment/infrastructure/persistence/jpa/PaymentJpaRepository.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.limito.payment.infrastructure.persistence.entity.PaymentEntity;
+import com.limito.payment.domain.entity.PaymentEntity;
 
 public interface PaymentJpaRepository extends JpaRepository<PaymentEntity, UUID> {
 }
